# swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Pokedex backend",
    "description": "This is a pokedex api with OAuth2 login",
    "version": "1.0.0",
    "contact": {
      "name": "Theo Declerck",
      "email": "theo.dclrck5@gmail.com"
    }
  },
  "host": "localhost:8000",
...
  "schemes": [
    "http"
  ],
  "paths": {
...
    "/trainer/register": {
      "post": {
        "tags": [
          "trainer"
        ],
        "summary": "Register a new Trainer.",
        "parameters": [
          {
            "name": "Trainer",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Trainer"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Trainer created"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/trainer/getTrainerInfo/self": {
      "get": {
        "tags": [
          "trainer"
        ],
        "summary": "Returns the current Trainer logged in.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/definitions/Trainer"
          },
          "404": {
            "description": "Trainer not found"
          }
        }
      }
    },
    "/trainer/getTrainerInfo/{trainerId}": {
      "get": {
        "tags": [
          "trainer"
        ],
        "summary": "Returns a Trainer by ID.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "trainerId",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/definitions/Trainer"
          },
          "404": {
            "description": "Trainer not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/trainer/updateLogin/self": {
      "patch": {
        "tags": [
          "trainer"
        ],
        "summary": "Update the current logged in Trainer login.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "body",
            "name": "login",
            "required": true,
            "schema": {
              "properties": {
                "login": {
                  "type": "string"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/definitions/Trainer"
          },
          "404": {
            "description": "Trainer not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/trainer/updateLogin/{trainerId}": {
      "patch": {
        "tags": [
          "trainer"
        ],
        "summary": "Update a Trainer login by ID.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "body",
            "name": "login",
            "required": true,
            "schema": {
              "properties": {
                "login": {
                  "type": "string"
                }
              }
            }
          },
          {
            "in": "path",
            "name": "trainerId",
            "type": "integer",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/definitions/Trainer"
          },
          "404": {
            "description": "Trainer not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/trainer/delete/self": {
      "delete": {
        "tags": [
          "trainer"
        ],
        "summary": "Delete the current logged in Trainer.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Trainer has been deleted"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/trainer/delete/{trainerId}": {
      "delete": {
        "tags": [
          "trainer"
        ],
        "summary": "Delete the current logged in Trainer.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "body",
            "name": "id",
            "required": true,
            "schema": {
              "properties": {
                "id": {
                  "type": "integer"
                }
              }
            }
          },
          {
            "in": "path",
            "name": "trainerId",
            "type": "string",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Trainer has been deleted"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/pokemon/add/self": {
      "post": {
        "tags": [
          "pokemon"
        ],
        "summary": "Add a Pokemon to the current logged in Trainer.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "body",
            "name": "Pokemon",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Pokemon"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Pokemon Added"
          },
          "404": {
            "description": "Trainer not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/pokemon/add/{trainerId}": {
      "post": {
        "tags": [
          "pokemon"
        ],
        "summary": "Add a Pokemon to a Trainer by ID.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "body",
            "name": "Pokemon",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Pokemon"
            }
          },
          {
            "in": "path",
            "name": "trainerId",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "Pokemon Added"
          },
          "404": {
            "description": "Trainer not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/pokemon/getPokemons/self": {
      "get": {
        "tags": [
          "pokemon"
        ],
        "summary": "Returns all Pokemons from the current logged in Trainer.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/definitions/Pokemon"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/pokemon/getPokemons/{id}": {
      "get": {
        "tags": [
          "pokemon"
        ],
        "summary": "Returns all Pokemons from a Trainer by ID.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/definitions/Pokemon"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/pokemon/release/self": {
      "delete": {
        "tags": [
          "pokemon"
        ],
        "summary": "Delete a Pokemon from the current logged in Trainer.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "body",
            "name": "id",
            "required": true,
            "schema": {
              "properties": {
                "id": {
                  "type": "integer"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Pokemon released"
          },
          "400": {
            "description": "This is not your pokemon"
          },
          "404": {
            "description": "Pokemon not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/pokemon/release/{trainerId}": {
      "delete": {
        "tags": [
          "pokemon"
        ],
        "summary": "Delete a Pokemon from a Trainer by ID.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "body",
            "name": "id",
            "required": true,
            "schema": {
              "properties": {
                "id": {
                  "type": "integer"
                }
              }
            }
          },
          {
            "in": "path",
            "name": "trainerId",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "Pokemon released"
          },
          "400": {
            "description": "The trainer does not own this pokemon"
          },
          "404": {
            "description": "Pokemon not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/pokemon/rename/self": {
      "patch": {
        "tags": [
          "pokemon"
        ],
        "summary": "Update a Pokemon name from the current logged in Trainer.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "body",
            "name": "name",
            "required": true,
            "schema": {
              "properties": {
                "name": {
                  "type": "string"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/definitions/Pokemon"
          },
          "400": {
            "description": "The trainer does not own this pokemon"
          },
          "404": {
            "description": "Pokemon not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/pokemon/rename/{trainerId}": {
      "patch": {
        "tags": [
          "pokemon"
        ],
        "summary": "Update a Pokemon name from a Trainer by ID.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "body",
            "name": "name",
            "required": true,
            "schema": {
              "properties": {
                "name": {
                  "type": "string"
                }
              }
            }
          },
          {
            "in": "path",
            "name": "trainerId",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/definitions/Pokemon"
          },
          "400": {
            "description": "The trainer does not own this pokemon"
          },
          "404": {
            "description": "Pokemon not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/pokemon/levelUp/{trainerId}": {
      "patch": {
        "tags": [
          "pokemon"
        ],
        "summary": "Update a Pokemon name from a Trainer by ID.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "body",
            "name": "level",
            "required": true,
            "schema": {
              "properties": {
                "level": {
                  "type": "integer"
                }
              }
            }
          },
          {
            "in": "path",
            "name": "trainerId",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/definitions/Pokemon"
          },
          "400": {
            "description": "The trainer does not own this pokemon"
          },
          "404": {
            "description": "Pokemon not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
...
  },
  "definitions": {
    "Trainer": {
      "properties": {
        "id": {
          "type": "integer"
        },
        "firstname": {
          "type": "string",
          "example": "valou"
        },
        "lastname": {
          "type": "string",
          "example": "depat"
        },
        "login": {
          "type": "string",
          "example": "ladep"
        },
        "age": {
          "type": "string",
          "example": "1999-10-10"
        }
      }
    },
    "Pokemon": {
      "properties": {
        "id": {
          "type": "integer"
        },
        "species": {
          "type": "string",
          "example": "carapuce"
        },
        "name": {
          "type": "string",
          "example": "karapute"
        },
        "level": {
          "type": "integer",
          "example": 5
        },
        "gender": {
          "properties": {
            "male": {
              "type": "string"
            },
            "female": {
              "type": "string"
            },
            "undefined": {
              "type": "string"
            }
          },
          "example": "male"
        },
        "size": {
          "type": "integer",
          "example": 35
        },
        "weight": {
          "type": "integer",
          "example": 6
        },
        "isShiny": {
          "type": "boolean",
          "example": true
        },
        "trainerId": {
          "type": "integer",
          "example": 1
        }
      }
    },
...
  },
  "securityDefinitions": {
    "bearerAuth": {
      "type": "apiKey",
      "name": "Token",
      "in": "header"
    }
  }
}